Don't count object converted into non-ref into clone limit
Previously, this:

    my $obj = bless([], "Foo");
    my $data = [$obj, $obj, $obj];
    my $cleanser = Data::Clean::JSON->new(Foo => ['replace_with_ref'], -circular=>[clone=>1]);
    my $cleaned = $cleanser->clone_and_clean($data);

will produce $cleaned as:

    ["Foo", "CIRCULAR", "CIRCULAR"]

because the second occurrence of Foo object is still counted into the clone
limit, even though the first object has been converted into a string. This
behavior is not desirable, for example when we convert many
DateTime/Time::Moment objects into Unix timestamps (ints).

This commit reorganizes how the cleaner code works. For every conversion, $ref
is updated to reflect the ref() of the new value. The circular check is moved to
after the object check, so if the current value of $ref is empty, we no longer
need to check for circular.